New logic

+++ src/main/java/may/code/api/store/repositories/AnswerRepository.java
@@ -3,5 +3,5 @@ package may.code.api.store.repositories;
 import may.code.api.store.entities.AnswerEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnswerRepository extends JpaRepository<AnswerEntity, Long> {
+public interface AnswerRepository extends JpaRepository<AnswerEntity, Integer> {
 }
