+++ src/main/java/may/code/api/store/repositories/PsychologistRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface PsychologistRepository extends JpaRepository<PsychologistEntity, Long> {
+public interface PsychologistRepository extends JpaRepository<PsychologistEntity, Integer> {
     Optional<PsychologistEntity> findTopByLoginAndPassword(@NonNull String login, @NonNull String password);
 }
