+++ src/main/java/may/code/api/store/repositories/QuestionRepository.java
@@ -3,5 +3,5 @@ package may.code.api.store.repositories;
 import may.code.api.store.entities.QuestionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestionRepository extends JpaRepository<QuestionEntity, Long> {
+public interface QuestionRepository extends JpaRepository<QuestionEntity, Integer> {
 }
